remove ssages sub-module
