+++ pysages/collective_variables/core.py
