+++ pysages/collective_variables/shape.py
