+++ pysages/methods/ann.py
@@ -3,46 +3,44 @@
 # See LICENSE.md and CONTRIBUTORS.md at https://github.com/SSAGESLabs/PySAGES
 
 from collections import namedtuple
-from jax.numpy import linalg
 from pysages.nn.models import mlp
 from pysages.nn.objectives import PartialRBObjective
 from pysages.nn.optimizers import LevenbergMaquardtBayes
 from pysages.nn.training import trainer
-from pysages.ssages.grids import build_indexer
+from pysages.grids import build_indexer
 
 from .core import NNSamplingMethod, generalize  # pylint: disable=relative-beyond-top-level
 
 import jax.numpy as np
 
 
-# ======== #
-#   FUNN   #
-# ======== #
+# ======= #
+#   ANN   #
+# ======= #
 
-class FUNNState(namedtuple(
-    "FUNNState",
+class ANNState(namedtuple(
+    "ANNState",
     (
         "bias",
         "nn",
         "hist",
-        "Fsum",
-        "F",
-        "Wp",
-        "Wp_",
+        "uhist",
+        "weight",
+        "weight_"
     )
 )):
     def __repr__(self):
         return repr("PySAGES " + type(self).__name__)
 
 
-class FUNN(NNSamplingMethod):
+class ANN(NNSamplingMethod):
     def __call__(self, snapshot, helpers):
         N = np.asarray(self.kwargs.get('N', 200))
-        return _funn(snapshot, self.cv, self.grid, self.topology, N, helpers)
+        return _ann(snapshot, self.cv, self.grid, self.topology, N, helpers)
 
 
-def _funn(snapshot, cv, grid, topology, N, helpers):
-    dt = snapshot.dt
+def _ann(snapshot, cv, grid, topology, N, helpers):
+    kBT = snapshot.kBT
     dims = grid.shape.size
     natoms = np.size(snapshot.positions, 0)
     get_grid_index = build_indexer(grid)
@@ -53,31 +51,29 @@ def _funn(snapshot, cv, grid, topology, N, helpers):
     def initialize():
         bias = np.zeros((natoms, dims))
         hist = np.zeros(grid.shape, dtype=np.uint32)
-        Fsum = np.zeros((*grid.shape, dims))
-        F = np.zeros(dims)
-        Wp = np.zeros(dims)
-        Wp_ = np.zeros(dims)
-        return FUNNState(bias, model.parameters, hist, Fsum, F, Wp, Wp_)
+        uhist = hist.copy()
+        weight = np.ones(1)
+        weight_ = np.zeros(1)
+        return ANNState(bias, model.parameters, hist, uhist, weight, weight_)
 
     def update(state, rs, vms, ids):
         # Compute the collective variable and its jacobian
         ξ, Jξ = cv(rs, indices(ids))
         #
-        θ = train(state.nn, state.Fsum / state.hist).θ
-        #
-        p = momenta(vms)
-        Wp = linalg.tensorsolve(Jξ @ Jξ.T, Jξ @ p)
-        dWp_dt = (1.5 * Wp - 2.0 * state.Wp + 0.5 * state.Wp_) / dt
+        θ = train(state.nn, state.bias).θ
+        F = model.apply(θ, ξ)
         #
         I_ξ = get_grid_index(ξ)
         N_ξ = state.hist[I_ξ] + 1
-        F_ξ = state.Fsum[I_ξ] + dWp_dt + state.F
+        H_ξ = state.weight_ * state.uhist[I_ξ] + \
+            state.weight * np.exp(state.bias[I_ξ] / kBT)
         hist = state.hist.at[I_ξ].set(N_ξ)
-        Fsum = state.Fsum.at[I_ξ].set(F_ξ)
-        F = model.apply(θ, ξ)  # F_ξ / np.maximum(N_ξ, N)
+        uhist = state.uhist.at[I_ξ].set(H_ξ)
+        bias = kBT * np.log(H_ξ)
+        bias = bias - bias.min()
         #
         bias = np.reshape(-Jξ.T @ F, state.bias.shape)
         #
-        return FUNNState(bias, θ, hist, Fsum, F, Wp, state.Wp)
+        return ANNState(bias, θ, hist, uhist, state.weight, state.weight_)
 
     return snapshot, initialize, generalize(update)
