+++ pysages/methods/harmonic_bias.py
