--- pysages/ssages/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (c) 2020-2021: PySAGES contributors
-# See LICENSE.md and CONTRIBUTORS.md at https://github.com/SSAGESLabs/PySAGES
-
-
-from . import collective_variables
-from . import methods
-from .grids import Grid
